feat: add api for Customer refer to issue #10

+++ src/main/java/com/example/sales_system/entity/tenant/Customer.java
@@ -39,6 +39,6 @@ public class Customer extends AbstractTimestampEntity {
     String province;
     String district;
 
-    Integer rewardPoint;
+    Long rewardPoint;
     String note;
 }
